a general sweep of the code to fix print formatting warnins that XCode 4.4 flag.
git-svn-id: svn://orca.physics.unc.edu/Orca@7202 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Cocoa Extensions/NSInvocation+Extensions.m
@@ -251,7 +251,7 @@
 	else if(!strncmp(theArg,"{_NSRa",6)){
 		NSRange theRange;
         [self getReturnValue:&theRange]; 
-		return [NSString stringWithFormat:@"@(%f,%f)",theRange.location,theRange.length];
+		return [NSString stringWithFormat:@"@(%f,%f)",(float)theRange.location,(float)theRange.length];
 	}
 	else if(!strncmp(theArg,"{_NSRe",6)){
 		NSRect theRect;
